I added a new function (draw_one_contour), now its possible to look at each contour retrieved from image (and that is within a given minimal shape diameter threshold).
As I was expecting, the order in each object shapes appear
in contour sequence (CvSeq) is *not* necessarily related to
spatial location of objects in image.

A programmatic way to infer the spatial location of shape
would work with its centroid coordinates. I think that
simply sorting with centroid as parameter would do the
trick (need probe it further).

To make this work, it was needed to change window widget
deallocation point (now, its outside of logical block of
batch mode X interactive). I really think that main app
code should be refactored, since its original design it
was hacked to accommodate new functionality.


git-svn-id: file:///home/adenilson/temp/wtf/repository@141 56e0a439-041b-0410-8790-949b9ad245cc